contributors/adobe/incoming/css3-regions/interactivity/keyboard/keyboard-events-region-onkeydown-onkeyup-correct-value.html

<!DOCTYPE HTML>
<html>
    <head>
        <title>CSS Regions: onkeydown events work inside region and outside region</title>
        <link rel="author" title="David Alcala" href="mailto:dalcala@adobe.com">
        <link rel="help" href="http://www.w3.org/TR/css3-regions/">
        <link rel="help" href="http://www.w3.org/TR/css3-regions/#the-flow-into-property">
        <link rel="help" href="http://www.w3.org/TR/css3-regions/#the-flow-from-property">
        <link rel="help" href="http://www.w3.org/TR/css3-regions/#relation-to-document-events">
        <meta name="assert" content="CSS regions module does not alter the normal processing of events in the document
         tree. Therefore, onkeydown and onkeyup events inside and outside a named flow should function correctly.">
        <meta name="flags" content="dom interact">
        <link href="support/css/regions-keyboard.css" rel="stylesheet" type="text/css">
        <link href="/resources/testharness.css" rel="stylesheet" type="text/css">
        <script src="/resources/testharness.js"></script>
    </head>
    <body>
        <p id="msg">
            keydown and keyup the "a" key in the text field in the black div, then keydown and keyup the "a" key in the text field in the orange div.
        </p>
        <div id="content">
            <div class="block" title="FAIL"></div>
            <div class="block" title="FAIL"></div>
            <div id="target-block" class="block">
                <form id="my-form">
                    <input type="text" id="textfield-inside" size="10">
                    <input type="button" id="myButton" value="Button">
                </form>
            </div>
        </div>
        <div class="region"></div>
        <div class="region"></div>
        <div class="region"></div>
        <br/>
        <div id="outside-region">This div is outside the region
            <form id="form-outside">
                <input type="text" id="textfield-outside" size="10">
                <input type="button" id="myButton" value="Button">
            </form>
        </div>
        <div id="log"></div>
    </body>
    <script type="text/javascript">
     
    //   runMouseTest();
        
        var testTimeout = 10000;
        setup({timeout: testTimeout});
        
        // This block is executed if running in WebKit's harness
        if (window.testRunner)
        {
        	testRunner.dumpAsText(false);
        }

        // Verify that CSS Regions are enabled in the browser.
        // Divs will be horizontal if Regions are enabled.
        // Divs will be vertical if Regions are not enabled.
        function checkLeftPosition(elemID) 
        {
            return document.getElementById(elemID).getBoundingClientRect().left;
        }
        
      //  function runMouseTest()
      //  {
        
        if(checkLeftPosition("target-block") != 0) {

            var test1 = async_test("onkeydown event fired in textfield inside region", {timeout: testTimeout});
                test1.step( function() {
                    var testTarget = document.getElementById("textfield-inside");
                    testTarget.onkeydown = test1.step_func( function(evt) {
                        /* Log assertion that the event got fired */
                        assert_true(true);

                        /* Verify the target inside the region has the correct ID */
                        test(function() { assert_equals(evt.target.id, "textfield-inside")}, "Target inside region has the correct ID");

                        test1.done();
                    });
                    //keyDown("textfield-inside");
                });

            var test2 = async_test("onkeyup event fired in textfield inside region", {timeout: testTimeout});
                test2.step( function() {
                    var testTarget = document.getElementById("textfield-inside");
                    testTarget.onkeyup = test2.step_func( function(evt) {

                        /* Verify the textfield inside the region has the correct value */
                        test(function() { assert_equals(evt.target.value, "a")}, "Textfield inside region has the correct value");

                        test2.done();
                    });
                    keyDown("textfield-inside");
                });
                
            var test3 = async_test("onkeydown event fired in textfield outside region", {timeout: testTimeout});
                test3.step( function() {
                    var testTarget = document.getElementById("outside-region");
                    testTarget.onkeydown = test3.step_func( function(evt) {
                        /* Log assertion that the event got fired */
                        assert_true(true);

                        /* Verify the target outside the region has the correct ID */
                        test(function() { assert_equals(evt.target.id, "textfield-outside")}, "Target outside region has the correct ID");

                        test3.done();
                    });
                    //keyDown("textfield-outside");
                });

            var test4 = async_test("onkeyup event fired in textfield outside region", {timeout: testTimeout});
                test4.step( function() {
                    var testTarget = document.getElementById("outside-region");
                    testTarget.onkeyup = test4.step_func( function(evt) {
                        /* Verify the textfield inside the region has the correct value */
                        test(function() { assert_equals(evt.target.value, "a")}, "Textfield outside region has the correct value");

                        test4.done();

                        //Update the message stating that tests are complete
                        var msg = document.getElementById("msg");

                        var complete = document.createElement("p");
                        complete.innerHTML = "Tests are complete. All results in the Details section below should PASS.";
                        complete.style.color = "blue";
                        msg.appendChild(complete);
                    });
                    keyDown("textfield-outside");
                });

         //   }  end of runMouseTest()

            function keyDown(block)
            {
            	if(window.testRunner)
               	{
            		var input = document.getElementById(block);
                    input.focus();
                    eventSender.keyDown('a');
            	}
            }

        } //end of checkLeftPosition test
        else { 
            test(function() {assert_true(false) }, "Regions are not enabled");
        }

    </script>
</html>
